feat!: use fixtures
Trying to work towards building dependency graphs in the benchmark
execution to avoid race conditions during build. For free, the wrapper
gains a much more expressive (and safe) way to wrap new benchmarks.
