+++ harness/utils.py
@@ -41,7 +41,9 @@ def extract_perf_values(socket, key, fd, group):
 def query_runtime(job):
     # check we have slurm
     if job.scheduler.registered_name != "slurm":
-        logger.info("Job does not use slurm. Cannot query better start / end time estimate.")
+        logger.info(
+            "Job does not use slurm. Cannot query better start / end time estimate."
+        )
         return None
 
     slurm_query = [
